pyproject.toml
[tool.poetry]
name = "chatgpt-asana-tasks"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "chatgpt_asana_tasks"}]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
types-requests = "^2.28.11.15"
asana = "^3.1.1"
openai = "^0.27.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
mypy = "^1.0.1"
black = "^23.1.0"
isort = "^5.12.0"
pytest = "^7.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
